webhost_app/requirements.txt
Flask==2.0.1
Flask-APScheduler==1.12.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
plotly==5.1.0
requests==2.26.0
python-dateutil==2.8.2
